Extract ToggleRejectedSongs from Playlist
